feat: use feedbacks from devices to update panels
In some scenarios, a button that's been added to the Navigaotr needs to change in response to
feedback or other values changing in a room or device plugin. The easiest way to accomplish that currently
is to use the existing feedback mechanism. In order for a device or room to be used with this mechanism, the device/room
MUST implement the `IHasFeedback` interface AND add keys to the feedbacks that are created and added to the FeedbackCollection.

The configuration for this feature allows for responding to all 3 types of feedbacks (bool, string, int/ushort), along with specifying either a new
name for the panel or a new color for the button. There is the possiblity of changing the location of a button, allowing for hiding/showing a button
or moving a button from a control panel to the home screen or any of the other locations that are valid.

Configuration looks like:
```json
{
...standard device stuff,
"properties":{
  ...config,
  "extensions":{
    ...extensionConfig,
    "panels": [
      {
        ...panelConfig,
        "panelFeedback": {
          "deviceKey": "room1",
          "feedbackKey: "myBoolFeedback",
          "propertyToChange": "color", //or name
 .        "feedbackEventType": "TypeBool",
          "truePropertyValue": "#000000" //has to be a hex color
          "falsePropertyValue": "#FF0000"
        },
        {
          "deviceKey": "room1",
          "feedbackKey": "myStringFeedback",
          "propertyToChange": "name",
          "feedbackEventType": "TypeString",
          "stringFeedbackPropertyValues": {
             "someValueFromFeedback": "Feedback is on",
             "someOtherValueFromFeedback" "Feedback is some other value",
          }
        },
        {
          "deviceKey": "room1",
          "feedbackKey": "myIntFeedback",
          "propertyToChange": "name",
          "feedbackEventType": "TypeInt",
          "intFeedbackPropertyValues": {
            1 : "some string here",
            45 : "some other string"
          }
        }
      }
    ]
  }
}
```