feat: customizable lockouts
The navigator now supports multiple customizable lockouts, along with the
existing room combination lockout. The existing lockout property in the configurations
will be used as the lockout when a navigator is in a room that is part of a combination scenario.

The customLockouts property will be used to enable reacting to the change in a boolean feedback
from any valid device in Essentials that implements the `IHasFeedback` interface along with the
feedback being defined with a key.