fix: add overload for SelfviewModeOn
This overload takes an enum for the monitor role to allow for moving the selfview between monitors. There is also an updated configuration value to allow for setting the default monitor role. If the default monitor role is set in configuration, the value will be sent when selfview is turned off.