feat: enhance web view handling
The ShowWebView method now checks if the target is OSD and directly sends the command to show the webview if it is.

If the target is NOT OSD, then the codec finds the first navigator who's parent codec has the same key, and calls that navigator's ShowWebViewOsd method.

The ShowWebViewOsd method will then parse the URL, figure out if it's a relative or absolute URL, and use the appropriate method to handle showing the webview.