fix: panel feedbacks now change to correct room
When a panel was configured to look for feedback from a room to change
an attribute, the wrong feedback could be subscribed to in combinable room scenarios. Now, when the room combination scenario changes, feedbacks are unsubscribed and resubscribed. Any subscriptions to rooms are moved to the new room key instead of the old room key.