fix: modify initialization for UI extensions
If extensions are configured on both a codec and a navigator, then the extensions configured on the codec will take precedence instead of using both sets. It may be beneficial to remove the option to configure Extensions directly on the codec device, as the Navigator device has now been added that accomplishes the same functionality.